Move writing current to io
